[UPD]有限体や多項式のimplを増やした。多項式の扱いとモジュラスへの対応が課題か。

+++ src/galois_field/mod.rs
@@ -1,3 +1,3 @@
-pub type Element = isize;
+pub type Element = u64;
 pub mod gf_context;
 pub mod gf_elm;
